+++ Template/Trigger/FullscreenTrigger.web.js
@@ -7,6 +7,10 @@
             function onFullScreen() {
                 var limit = parameters.get('triggerLimit', 1);
 
+                if (limit) {
+                    limit = parseInt(limit, 10);
+                }
+
                 if (limit && limit <= numTriggers) {
                     return;
                 }