+++ angularjs/manageVariable/edit.controller.js
@@ -38,7 +38,7 @@
         });
 
         // needed for suggestNameForType() to make sure it is aware of all names
-        this.model.fetchVariablesIfNotLoaded();
+        this.model.fetchVariablesIfNotLoaded(this.idContainer, this.idContainerVersion);
 
         function enrichTemplateType(template)
         {