+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  * Use Custom Template
  * Publish Live Contaier
 * New API method `TagManager.getAvailableEnvironmentsWithPublishCapability`
+* Custom HTML tag supports assigning JavaScript variables, CSS styles and HTML content
 
 0.1.1
 * Fixed CustomEventsTrigger may miss an event when pushed before tag manager is loaded